requirements.txt
# pip install -r requirements.txt

# ------------------------------------------
# 🌐 Django и сопутствующие пакеты
# ------------------------------------------
Django>=5.2,<6.0         # Веб-фреймворк, основа всего проекта

# ------------------------------------------
# 📊 Обработка данных
# ------------------------------------------
pandas>=2.0              # Построение DataFrame из EntryValue и чтение Excel
numpy>=1.24              # Поддержка числовых операций, требуется pandas

# ------------------------------------------
# 🤖 Машинное обучение (линейная модель)
# ------------------------------------------
scikit-learn>=1.2        # Используется для LinearRegression (базовая модель)
joblib>=1.2              # Сохраняем и загружаем модели в формате .pkl

# ------------------------------------------
# 📥 Загрузка Excel-файлов через админку
# ------------------------------------------
openpyxl>=3.1            # Движок для чтения .xlsx в pandas.read_excel()
python-slugify>=8.0  # 🔤 Преобразует строку (например, "Тошнота сильная!") в slug (например, "toshnota-silnaya")
                     # Используется для генерации уникального ключа `key` на основе `name_ru`, если его нет.
                     # Нужен при автоматическом создании новых параметров из заголовков Excel-файла.


# ------------------------------------------
# 🛠️ Разное (опционально, для удобства)
# ------------------------------------------
ipython                  # Удобный интерактивный shell (manage.py shell_plus)
python-dotenv            # Для подключения .env-файлов, если используешь их
